Tidy up property handling and make integers long :)

+++ xsd2vdm/src/main/java/values/SimpleValue.java
@@ -57,10 +57,10 @@ public class SimpleValue extends VDMValue
 		this.value = quote.toString();
 	}
 
-	public SimpleValue(Type type, Locator locator, Integer value)
+	public SimpleValue(Type type, Locator locator, Long value)
 	{
 		super(type, locator);
-		this.value = Integer.toString(value);
+		this.value = Long.toString(value);
 	}
 
 	public SimpleValue(BasicType type, Locator locator, Double real)
